show path to and version of solc
This adds a pytest_report_header hook, which print the path to the solc
executable and tries to determine the version by running solc --version.

This is copied from the conftest.py file of the contracts repository.

This information is currently shown whenever the plugin is installed. The plan
is to let consumers of the plugin explicitly declare the dependency on the
plugin.

see https://github.com/trustlines-network/contract-deploy-tools/issues/20