do not fail if all gas has been used
wait_for_successful_transaction_receipt did check if all gas up to the gas limit
had been used and raised an error in that case. The assumption was that a
failing transaction would use up all gas.

We now have the status field inside the transaction receipt and use that to
check if the transaction was successful.

When using infura, we end up with successful transactions that use up all the
gas. These should succeed now.