Add workaround for infura's broken getTransactionReceipt
When using infura, web3.eth.waitForTransactionReceipt returns a receipt with a
block number which is None, which probably also means that the transaction
hasn't made it to the chain. This is clearly not what we want.

This commit adds an adapted wait_for_transaction_receipt function, which we use
instead of web3.eth.getTransactionReceipt. It will make sure that we've also
have seen a block hash.

It looks like this may fix
https://github.com/trustlines-protocol/blockchain/issues/70