pin the evm target version to byzantium
This adds evm_version as an argument to compile.compile_project and uses
"byzantium" as default.

solidity changed the default target version between 0.5.3 and 0.5.7. Our test
infrastructure cannot handle the new version at the moment, so we effectively
pin the version to "byzantium"

We could/should add arguments to specify the target version to the command line
tool later.